+++ include/basic/transaction_list.h
@@ -16,18 +16,18 @@ void transaction_list_set(struct transaction_list_st *, const struct transaction
 void transaction_list_copy(struct transaction_list_st *, const struct transaction_list_st *);
 
 void transaction_list_clear(struct transaction_list_st *);
-int transaction_list_cmp(const struct transaction_list_st *, const struct transaction_list_st *);
+int8_t transaction_list_cmp(const struct transaction_list_st *, const struct transaction_list_st *);
 
 // Data Methods
 void transaction_list_data_init(struct transaction_list_st *);
 void transaction_list_data_free(struct transaction_list_st *);
 
 // Cmp Methods
-int transaction_list_is_null(const struct transaction_list_st *);
+int8_t transaction_list_is_null(const struct transaction_list_st *);
 void transaction_list_resize(struct transaction_list_st *, size_t);
 
 // TLV Methods
-int transaction_list_set_tlv(struct transaction_list_st *, const struct string_st *);
+int8_t transaction_list_set_tlv(struct transaction_list_st *, const struct string_st *);
 void transaction_list_get_tlv(const struct transaction_list_st *, struct string_st *);
 
 // Attrib Methods
