+++ include/smart/bytecode_list.h
@@ -7,7 +7,7 @@ struct bytecode_list_st{
     struct bytecode_st** bytecodes;
     size_t size;
     size_t max_size;
-    int type;
+    int8_t type;
 };
 
 void bytecode_list_set(struct bytecode_list_st *, const struct bytecode_list_st *);
@@ -23,7 +23,7 @@ struct bytecode_st *bytecode_list_last(struct bytecode_list_st *);
 struct bytecode_st *bytecode_list_pop(struct bytecode_list_st *res);
 
 // TLV Methods
-int bytecode_list_set_tlv(struct bytecode_list_st *, const struct string_st *);
+int8_t bytecode_list_set_tlv(struct bytecode_list_st *, const struct string_st *);
 void bytecode_list_get_tlv(const struct bytecode_list_st *, struct string_st *);
 
 void print_bytecode_list(const struct bytecode_list_st *, int);
