+++ include/smart/closure_list.h
@@ -7,7 +7,7 @@ struct closure_list_st{
     struct closure_st **closures;
     size_t size;
     size_t max_size;
-    int type;
+    int8_t type;
 };
 
 void closure_list_set(struct closure_list_st *, const struct closure_list_st *);
@@ -22,7 +22,7 @@ size_t closure_list_add_new(struct closure_list_st *);
 struct closure_st *closure_list_last(struct closure_list_st *);
 
 // TLV Methods
-int closure_list_set_tlv(struct closure_list_st *, const struct string_st *);
+int8_t closure_list_set_tlv(struct closure_list_st *, const struct string_st *);
 void closure_list_get_tlv(const struct closure_list_st *, struct string_st *);
 
 void print_closure_list(const struct closure_list_st *, int);
